2.2 USERDAOTEST 개선 및 패키지 리팩토링
