+++ notebook/yuuuunmi/2. 테스트.md
@@ -87,53 +87,62 @@
 
 **수동 확인 작업의 번거로움** 데이터의 준비를 모두 자동으로 진행하도록 만들어졌지만, 여전히 사람의 눈으로 확인하는 과정이 필요하다.  테스트의 결과를 확인하는 일은 사람의 책임이므로 완전히 자동으로 테스트되는 방법이라고 말할 수가 없다. 이는 불편하고, 실수를 할 가능성도 있다.
 
-**실행 작업의 번거로움** 아무리 간단한 `main()` 메소드라고 하더라도 매번 그것을 실행하는 것은 번거롭다. DAO가 수백개가 된다면 전체 기능을 테스트 해보기 위해서는 `main()`메소드를 수백번 실행하는 수고가 필요하다. 또한 눈으로 결과를 확인하고 결과를 정리하려면 이것 또한 큰 작업이 된다. 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
+**실행 작업의 번거로움** 아무리 간단한 `main()` 메소드라고 하더라도 매번 그것을 실행하는 것은 번거롭다. DAO가 수백개가 된다면 전체 기능을 테스트 해보기 위해서는 `main()`메소드를 수백번 실행하는 수고가 필요하다. 또한 눈으로 결과를 확인하고 결과를 정리하려면 이것 또한 큰 작업이 된다.
 
 
 # UserDaoTest 개선
+
+## 2.2.1 테스트 검증의 자동화
+
+모든 테스트는 성공과 실패의 두 가지 결과를 가질 수 있다.  
+
+테스트의 실패는
+**테스트 에러** : 테스트가 진행되는 동안에 에러가 발생해서 실패하는 경우
+**테스트 실패** : 기대한 것과 다르게 나오는 경우
+두 가지 결과를 가질 수 있다.
+
+테스트 에러는 콘솔에 에러 메세지와 긴 호출 스택 정보가 출력되기 때문에 쉽게 확인이 가능하다.
+하지만 테스트가 실패하는 것은 별도의 확인 작업과 그 결과가 있어야만 알 수 있다.
+
+`UserDaoTest`는 `add()`에 전달한 `User` 오브젝트에 담긴 사용자 정보와 `get()`을 통해 다시 DB에서 가져온 `User` 오브젝트의 정보가 서로 정확히 일치하는가를 확인한다.
+기존에는 단순히 콘솔에 출력하기만 했는데, 기대한 결과와 달라서 실패했을 경우에는 코드에서 직접 결과를 확인하여 "테스트 실패"라는 메세지를 출력하도록 만들어보자. 그리고 모든 확인 작업을 통과하면 그때는 "테스트 성공"이라고 출력하도록 한다.
+
+이렇게 하면 테스트의 수행과 테스트 값의 적용, 그리고 결과를 검증하는 것까지 모두 자동화한 것이다. 할 일은 마지막 출력 메시지가 "테스트 성공"이라고 나오는지 확인하는 것뿐이다.
+
+이 테스트는 `UserDao`의 기능이 정상적으로 동작하는지 언제든지 손쉽게 확인할 수 있게 해주며, 코드의 동작에 영향을 미칠 수 있는 어떤 변화라도 생기면 언제든지 다시 실행해볼 수 있다.
+
+특히 포괄적인 테스트(Comprehensive Test, 만들어진 코드의 기능을 모두 점검)를 만들면서부터는, 과감한 수정을 한 후라도 테스트를 모두 돌려보고 나면 안심이 된다. 혹은 테스트를 통해 그 변경에 영향을 받는 부분이 정확히 확인된다면 빠르게 조치를 취할 수 있다.
+
+빠르게 실행 가능하고 스스로 테스트 수행과 기대하는 결과에 대한 확인까지 해주는 코드로 된 자동화된 테스트를 만들어두면
+- 코드 수정 시 마음의 평안을 얻고
+- 코드에 대해 자신감을 가질 수 있다.
+
+
+
+## 2.2.2 테스트의 효율적인 수행과 결과 관리
+
+좀 더 편리하게 테스트를 수행하고 결과를 확인하려면 단순 `main()` 메소드로는 한계가 있다.
+- 일정한 패턴을 가진 테스트를 만들 수 있고
+- 많은 테스트를 간단히 실행시킬 수 있으며
+- 테스트 결과를 종합해서 볼 수 있고,
+- 테스트가 실패한 곳을 빠르게 찾을 수 있는 기능
+을 갖춘 테스트 지원 도구와 그에 맞는 테스트 작성 방법이 필요하다.
+`main()` 메소드를 이용한 방법만으로는 애플리케이션 규모가 커지고 테스트 개수가 많아지면 테스트를 수행하는 일이 점점 부담이 될 것이다. 
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
