+++ source/yuuuunmi/src/applicationContext.xml
@@ -2,7 +2,8 @@
 <beans xmlns="http://www.springframework.org/schema/beans"
        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
        xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
-    <bean id="userDao" class="springbook.user.domain.UserDao">
+
+    <bean id="userDao" class="springbook.user.dao.UserDao">
         <property name="dataSource" ref="dataSource"/>
     </bean>
