+++ source/yuuuunmi/src/springbook/user/dao/ConnectionMaker.java
@@ -1,4 +1,4 @@
-package springbook.user.domain;
+package springbook.user.dao;
 
 import java.sql.Connection;
 import java.sql.SQLException;
