+++ source/yuuuunmi/src/springbook/user/dao/DaoFactory.java
@@ -1,4 +1,4 @@
-package springbook.user.domain;
+package springbook.user.dao;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
