--- source/yuuuunmi/src/springbook/user/domain/UserDaoTest.java
@@ -1,35 +1,0 @@
-package springbook.user.domain;
-
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.support.GenericXmlApplicationContext;
-
-import java.sql.SQLException;
-
-public class UserDaoTest {
-    public static void main(String[] args) throws SQLException {
-
-        ApplicationContext context = new GenericXmlApplicationContext("applicationContext.xml");
-
-        UserDao dao = context.getBean("userDao", UserDao.class);
-
-        User user = new User();
-        user.setId("dbsal");
-        user.setName("윤미");
-        user.setPassword("yyyyy");
-
-        dao.add(user);
-
-        System.out.println(user.getId() + " 등록 성공 ! ");
-
-        User user2 = dao.get(user.getId());
-
-        if (!user.getName().equals(user2.getName())) {
-            System.out.println("테스트 실패 (name)");
-        } else if (!user.getPassword().equals(user2.getPassword())) {
-            System.out.println("테스트 실패 (password)");
-        } else {
-            System.out.println("조회 테스트 성공");
-        }
-
-    }
-}
